requirements.txt
arrow==1.3.0
asgiref==3.6.0
asn1crypto==1.5.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bitarray==2.9.3
black==25.1.0
boto3==1.26.104
botocore==1.29.104
Brotli==1.0.9
cattrs==24.1.2
certifi==2022.12.7
certvalidator==0.11.1
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.1
cssselect2==0.7.0
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.2
django-admin-charts==1.4.0
django-admin-logs==1.3.0
django-appconf==1.0.6
django-bower==5.2.0
django-compressor==4.5.1
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-dotenv==1.4.2
django-endesive==0.2.0
django-htmlmin==0.11.0
django-htmx==1.14.0
django-import-export==3.1.0
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-schema-graph==3.1.0
django-sendgrid-v5==1.2.2
django-simple-history==3.3.0
django-simple-pwa==2.2.1
django-storages==1.13.2
django-tailwind==3.8.0
django-watchman==1.3.0
endesive==2.17.3
et-xmlfile==1.1.0
filelock==3.12.0
fonttools==4.39.3
greenlet==3.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
mypy_extensions==1.1.0
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==25.0
paramiko==3.5.0
pathspec==0.12.1
pdfreader==0.1.15
pillow==11.1.0
pipenv==2023.4.29
platformdirs==3.5.0
psycogreen==1.0.2
psycopg2-binary==2.9.10
pycparser==2.21
pycryptodome==3.21.0
pydyf==0.6.0
Pygments==2.18.0
PyKCS11==1.5.16
PyMuPDF==1.24.12
PyNaCl==1.5.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pyphen==0.14.0
python-dateutil==2.8.2
python-http-client==3.3.7
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2023.3
PyYAML==5.3.1
rcssmin==1.1.2
reportlab==4.2.5
requests==2.28.2
rich==13.9.4
rjsmin==1.2.2
s3transfer==0.6.0
sendgrid==6.10.0
setuptools==78.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tablib==3.4.0
text-unidecode==1.3
tinycss2==1.2.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
virtualenv-clone==0.5.7
weasyprint==58.1
webencodings==0.5.1
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
zope.event==5.0
zope.interface==6.0
zopfli==0.2.2
